moved the method PMA_getHtmlForDisplayIndexes to the index.lib.php file
